% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyexcel.R
\name{fancy_excel}
\alias{fancy_excel}
\title{Title}
\usage{
fancy_excel(data, title = "Title")
}
\arguments{
\item{data}{A dataframe}

\item{title}{Title for the sheet to be added}
}
\value{

}
\description{
Title
}
\examples{
df <- dataset::swiss
fancy_excel(head(df), title = "My fancy title")
}
