% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyexcel.R
\name{fancy_excel}
\alias{fancy_excel}
\title{fancy_excel}
\usage{
fancy_excel(data, title = "Title", source = "Quelle:")
}
\arguments{
\item{data}{A dataframe}

\item{title}{Title for the sheet to be added}

\item{source}{Quellenangabe hinzufuegen}
}
\value{

}
\description{
Let's you write a dataframe to excel
}
\examples{
df <- datasets::swiss
fancy_excel(head(df), title = "My fancy title")
}
