% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyexcel.R
\name{fancy_excel}
\alias{fancy_excel}
\title{Add data to an Excel file}
\usage{
fancy_excel(data, title = "Title")
}
\arguments{
\item{x}{A data frame}
}
\description{
Add data to an Excel file
}
\examples{
df <- dataset::swiss
fancy_excel(head(df))


}
